move models to init
